fix(operations): widen duration accounting to bigint
Problem: operations_log duration_ms and its runtime projections used i32, so long-lived operations saturated or failed during SQLx type checking after the schema widened.\n\nWhat changed: converge existing operation rows to BIGINT, update SQL functions and repository/RPC/view types, and preserve large replay, tombstone, content, media, and email operation durations without i32 clamping.\n\nCompatibility/migration: schema apply converts the existing integer column with an explicit bigint cast before runtime queries use the new type.

+++ crate/sinex-db/src/repositories/replay.rs
@@ -301,7 +301,7 @@ impl ReplayRepository<'_> {
         meta_json: JsonValue,
         status: &str,
         msg: &str,
-        duration_ms: Option<i32>,
+        duration_ms: Option<i64>,
     ) -> Result<()> {
         sqlx::query!(
             r#"
